requirements.txt
beautifulsoup4==4.9.3
ipython==7.25.0
matplotlib==3.4.2
notebook==6.4.0
python-dotenv==0.18.0
requests==2.25.1
scikit-image==0.18.2
selenium==3.141.0
